test: features tests moved to existing controller tests
