+++ tests/Controllers/HomeControllerTest.php
@@ -4,6 +4,7 @@ namespace Tests\Controllers;
 
 use App\Models\TimeEntry;
 use App\Repositories\DashboardRepository;
+use App\Repositories\UserRepository;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Mockery\MockInterface;
 use Tests\TestCase;
@@ -15,11 +16,13 @@ class HomeControllerTest extends TestCase
     /** @var MockInterface */
     protected $dashboardRepository;
 
+    /** @var MockInterface */
+    private $userRepository;
+
     public function setUp(): void
     {
         parent::setUp();
         $this->signInWithDefaultAdminUser();
-        $this->withHeaders(['X-Requested-With' => 'XMLHttpRequest']);
     }
 
     private function mockRepository()
@@ -28,10 +31,29 @@ class HomeControllerTest extends TestCase
         app()->instance(DashboardRepository::class, $this->dashboardRepository);
     }
 
-    public function tearDown(): void
+    private function mockUserRepo()
     {
-        parent::tearDown();
-        \Mockery::close();
+        $this->userRepository = \Mockery::mock(UserRepository::class);
+        app()->instance(UserRepository::class, $this->userRepository);
+    }
+
+    /** @test */
+    public function it_shows_dashboard()
+    {
+        $this->mockUserRepo();
+
+        $mockedResponse = [['id' => 1, 'name' => 'Dummy User']];
+        $this->userRepository->expects('getUserList')
+            ->andReturn($mockedResponse);
+
+        $response = $this->get(route('home'));
+
+        $response->assertStatus(200)
+            ->assertViewIs('dashboard.index')
+            ->assertViewHas('users', $mockedResponse)
+            ->assertSeeText('Dashboard')
+            ->assertSeeText('Custom Report')
+            ->assertSeeText('Daily Work Report');
     }
 
     /** @test */
